+++ utils/get_VIIRS_SCA.py
@@ -62,7 +62,7 @@ else:
     
 #set multiprocessing limits
 CPUS = len(os.sched_getaffinity(0))
-CPUS = (CPUS/2)-2
+CPUS = int((CPUS/2)-2)
 
 
 def get_VIIRS_from_AWS():
